// stage-pfe-front/src/cni-api/src/models/formations-dto.ts
/* tslint:disable */
export interface FormationsDto {
  description?: string;
  duree?: number;
  id?: number;
  niveau?: string;
  objectifsFormation?: string;
  photo?: string;
  placesMax?: number;
  prerequis?: string;
  prix?: number;
  programmeDetaille?: string;
  statut?: string;
  themeId?: number;
  titre?: string;
}